agency/requirements3.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
form==0.0.1
phonenumbers==8.12.53
Pillow==9.2.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
